Two player Snake Game with User and Snake AI
The green snake is controlled by the user. Please press 'up', 'down', 'left', and 'right' on the keyboard to move the snake in the respective direction. The competitor is the Snake AI that uses adversarial search method named as Minimax with the inclusion of alpha beta pruning for more efficient performance.